Improvements include search with enter,UI improved for video,search and recommendation pages and skeleton load added
