+++ package.json
@@ -14,11 +14,13 @@
     "moment": "^2.29.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-loading": "^2.0.3",
     "react-loading-skeleton": "^2.2.0",
     "react-player": "^2.9.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
-    "react-youtube": "^7.13.1"
+    "react-youtube": "^7.13.1",
+    "skeleton": "^0.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
