+++ src/Components/ChannelRow/ChannelRow.css
@@ -1,7 +1,8 @@
 .channelRow{
     display: flex;
     align-items: center;
-    width: 75%;
+    width: 100%;
+    padding-right:20px
 }
 .channelRow_text{
     display: flex;    
@@ -26,4 +27,9 @@
     align-items: center;
     justify-content: space-between;
     color: black;
+}
+.channelRow:hover{
+box-shadow: 0 1px 4px 0 rgba(0,0,0,0.1);
+  transition: 0.3s;
+  border-radius: 5px;
 }
