+++ src/Components/Recommended/Recommended.css
@@ -20,8 +20,3 @@
     margin-left: 5%;
 }
 
-.loading{
-    position: fixed; /* or absolute */
-    top: 50%;
-    left: 50%;
-}
