+++ src/Components/VideoCard/VideoCard.css
@@ -1,13 +1,19 @@
 .videocard {
   margin-bottom: 50px;
-  width: 270px;
-  padding-right: 50px;
+  width: 280px;
+  padding-right: 5px;
+}
+.videocard:hover{
+box-shadow: 0 1px 4px 0 rgba(0,0,0,0.1);
+  transition: 0.3s;
+  border-radius: 5px;
 }
 .videocard_thumbnail {
   height: 170px;
-  width: 280px;
+  width: 278px;
   position: relative;
   display: inline-block;
+   border-radius: 5px 5px 0 0;
 }
 
 .videocard_info {
