+++ src/Components/VideoRow/VideoRow.css
@@ -1,19 +1,21 @@
-.videoRow{
+ .videoRow{
     display: flex;
-    margin-bottom: 20px;
+    margin-bottom: 10px;
     margin-top: 20px;
-    margin-left: 5%;
-    object-fit: contain;
-    min-width: 645px;
-    max-width: 871px;
+    width:100%;   
 }
 
+.videoRow:hover{
+ box-shadow: 0 2px 4px 0 rgba(0,0,0,0.2);
+  transition: 0.3s;
+  border-radius: 5px;
+}
 
-.videoRow>img{
+.videoRow>img{  
     object-fit: contain;
     width: 246px;
     height: 150px;
-
+    border-radius: 20px 20px 20px 20px;
 }
 
 
@@ -21,12 +23,14 @@
     margin-left: 14px;
     color:black;
     font-size: 12px;
+  
 }
 
 .videoRow_channel{
     display:flex;
     margin-top: 15px;
 }
+
 .videoRow_channelName{
     margin-left: 5px;
     font-size: 14px;   
@@ -44,20 +48,11 @@
     color: black;  
     overflow-wrap:break-word;
     word-break: break-all;
-    width:320px;
+    /* width:80%; */
+
    
 }
-.videoRow_description{
-    font-size: 12px;
-    color: gray;
-    overflow:hidden;
-    white-space:nowrap;
-    width:310px;
-    font-weight: 500;
-    margin-top: 20px;
-    text-overflow: ellipsis;
-    
-}
+
 
 .videoRow_overlay{
     color:black;
@@ -66,10 +61,5 @@
     margin-top: 3px;
     display: flex;
     flex-direction: column;
-    
-    
-    
-}
-.videoRow_moreicon{
-    color:black;  
-}
+
+}
